Updated tests for Proxy
